Updated to work with HTML <select>
This allows the user to search the database by first name, last name, email or occupation giving the search bar more accuracy. The users.json file allows the the users password, name, id and email to be saved to a file instead of just a blank array each time. This makes it so when the application is restarted you don't have to register a new account each time.